fix : user nickname 변경에러수정

See merge request s09-final/S09P31B303!99

+++ greenstep/src/main/java/com/mm/greenstep/domain/achieve/dto/response/AchieveDetailResDto.java
@@ -12,7 +12,7 @@ public class AchieveDetailResDto {
     private LocalDate createdAt;
     private Integer achieveTrashAmount; // 해당 업적 쓰레기 량
     private Double achieveTravelRange; // 해당 업적 이동거리
-    private Double achieveTravelTime; // 해당 업적 이동시간
+    private Long achieveTravelTime; // 해당 업적 이동시간
     private Integer achievePloggingCount; // 해당 업적 플로깅 횟수
     private Integer myTrashAmount; // 나의 총 쓰레기 량
     private Double myTravelRange; // 나의 총 이동거리
